+++ core/src/main/java/io/grpc/internal/GrpcAttributes.java
@@ -54,10 +54,9 @@ public final class GrpcAttributes {
    * The security level of the transport.  If it's not present, {@link SecurityLevel#NONE} should be
    * assumed.
    */
-  @SuppressWarnings("deprecation")
   @Grpc.TransportAttr
   public static final Attributes.Key<SecurityLevel> ATTR_SECURITY_LEVEL =
-      io.grpc.CallCredentials.ATTR_SECURITY_LEVEL;
+      Attributes.Key.create("io.grpc.internal.GrpcAttributes.securityLevel");
 
   private GrpcAttributes() {}
 }
